% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tindex.R
\name{tindex}
\alias{tindex}
\title{Simple function to create intra-group ranking and count variables.}
\usage{
tindex(df, grp, sortvar = NULL, output = "index", vector = T,
  name = NULL, desc = F)
}
\arguments{
\item{df}{An R \code{\link{data.frame}}}

\item{grp}{A grouping variable, such as a subject number, a classroom, a county, etc.}

\item{sortvar}{Sorts the data within grp (see above) by an additional variable (e.g., mpg within cyl in matcars). If no variable df is specified, data are sorted by grp.}

\item{output}{What type of output is desired, a within group "index" (the default) or a "count" within grp?}

\item{vector}{When true (the default), an index or count vector is returned which matches the sorting of df. When false, a data.frame sorted by grp and, optionally, sortvar, is returned with an additional count or index or count column.}

\item{name}{Ovveride the default column names returned when vector is false ("index" or "count").

Examples:

Build a count variable by number of cylinders, returning a vector:
mtcars$cylinder_count <- tindex(mtcars, cyl, output="count")

Build a mile per gallon index by number of cylinders, returning a sorted data frame with variable "cyl_rank":
ranked <- tindex(mtcars, cyl, sortvar=mpg, output="index", vector=F, name="cyl_rank")

Same as the previous example but reverses the order of "cyl_rank":
ranked <- tindex(mtcars, cyl, sortvar=mpg, output="index", vector=F, name="cyl_rank", desc=T)}

\item{desc}{When true, sortvar becomes a descending sort within grp.}
}
\description{
This function creates an index or count variable by group, optionally sorting on an additional parameter of interest.
For example, one may create an index/ranking of income within some group, e.g. the median income of census tracts within counties of a state. Or, one may create
a static count within some grouping variable, e.g. class size within a dataframe of test scores.
}

